user registration

+++ src/utils/cloudinary.js
@@ -1,5 +1,9 @@
-import {v2 as cloudinary} from 'cloudinary'
+
+import {v2 as cloudinary} from "cloudinary"
 import fs from 'fs'
+import dotenv from "dotenv";
+
+dotenv.config();  // <-- IMPORTANT
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -8,7 +12,7 @@ cloudinary.config({
 });
 
 
- const uploadCloudinary= async (localfilePath)=>{
+ const uploadOnCloudinary= async (localfilePath)=>{
 
      try{
            if(!localfilePath) return null
@@ -21,18 +25,23 @@ cloudinary.config({
                   //  file has been uploaded succesfully
                   
                   console.log("File uploaded on cloudinary successfully",response.url);
+                  fs.unlinkSync(localfilePath); // remove the locally saved file
                   return response ;
 
      }  catch(error){
 
-      fs.unlinkSync(localfilePath); // remove the locally saved file as the upload operation got failed
-              return null;
+     console.log("Cloudinary upload error:",error.message);
 
+   // Remove locally saved file
+        if (fs.existsSync(localfilePath)) {
+            fs.unlinkSync(localfilePath);
+        }
+
+        return null;
      }
 
  }
-
- export { uploadCloudinary }
+ export { uploadOnCloudinary }
